requirements.txt
google-generativeai >=0.8.2
ratelimit >=2.2.1
